+++ docs/CHANGELOG.md
@@ -1,12 +1,23 @@
-1.0.15
+Changelog
+=========
+
+
+1.0.16 (October, 16, 2019)
+-----------------------------------------
+- Enh: Update Twemoji to v12.1.3
+
+
+1.0.15 
 -----------------------------------------
 - Fix: emoji path configuration ignored
 - Enh: Added submit shortcut (default Ctrl+S)
 - Fix https://github.com/humhub/humhub/issues/3441 Richtext code block after list item not working
 
+
 1.0.14 (October 23, 2018)
 -----------------------------------------
 - Fix: Downgraded to prosemirror-dropcursor v 1.0.1
+
 
 1.0.13 (October 23, 2018)
 -----------------------------------------
@@ -18,6 +29,7 @@
 -----------------------------------------
 - Fix: state.storedMarks are cleaned by mention plugin.
 - Fix: `MarkdownEditor.clear()` does not cleanup $stage cached node
+
 
 1.0.11 (October 10, 2018)
 -----------------------------------------
@@ -32,6 +44,7 @@
 - Fix https://github.com/humhub/humhub/issues/3287 inline code block encoding
 - Enh: Added `preventMarks` on MarkSpec (currently only supported by the link menu item)
 
+
 1.0.9 (August 30, 2018)
 -----------------------------------------
 - Enh: Use of file-guid for uploaded images and files
@@ -44,18 +57,22 @@
 - Enh: Readded `anchor` plugin
 - Enh: Added prosemirror core to exports
 
+
 1.0.8 (July 24, 2018)
 -----------------------------------------
 - Chng: Use of relative upload url
+
 
 1.0.6 (July 30, 2018)
 -----------------------------------------
 - Fix: `context.getPluginOption()` default always overwrites actual option
 - Chng: Added `code` mark to default visible elements
 
+
 1.0.5 (July 18, 2018)
 -----------------------------------------
 - Fix: File upload errors not handled
+
 
 1.0.4 (July 17, 2018)
 -----------------------------------------